+++ src/AbcdSeq.cpp
@@ -36,6 +36,7 @@ struct AbcdSeq : Module,QuantizeUtils {
 		RND_VELS_INPUT,
 		LENGTH_INPUT,
 		RND_LENGTHS_INPUT = LENGTH_INPUT + 4,
+		SEED_INPUT,
 		NUM_INPUTS
 	};
 	enum OutputIds {
@@ -155,6 +156,7 @@ struct AbcdSeq : Module,QuantizeUtils {
 		configInput(RND_CV_INPUT, "Random Notes");
 		configInput(RND_TEXT_INPUT, "Random Text");
 		configInput(RND_LENGTHS_INPUT, "Random Lengths");
+		configInput(SEED_INPUT, "Seed input for repeatable randomization");
 		
 		configOutput(GATES_OUTPUT, "Gate");
 		configOutput(CV_OUTPUT, "CV");
@@ -525,7 +527,14 @@ struct AbcdSeq : Module,QuantizeUtils {
 // STEP
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 void AbcdSeq::process(const ProcessArgs &args) {
-    if(!initialRowSet){
+	if(inputs[SEED_INPUT].isConnected()) {
+		float f = clamp(inputs[SEED_INPUT].getVoltage(),0.f,10.f);
+		if (f != 0.f) {
+			auto seed = static_cast<uint64_t>(f*static_cast<float>(std::numeric_limits<uint32_t>::max()));
+			random::local().seed(seed,seed/7);
+		}
+	}
+	if(!initialRowSet){
         row = getRowForChar(text[charIdx]);
         initialRowSet = true;
     }
@@ -745,6 +754,7 @@ AbcdSeqWidget::AbcdSeqWidget(AbcdSeq *module) {
 	noteKnob->connectLabel(noteLabel, module);
 	addChild(noteLabel);
 	addParam(noteKnob);
+	addInput(createInput<TinyPJ301MPort>(Vec(47, 355), module, AbcdSeq::SEED_INPUT));
 	addInput(createInput<TinyPJ301MPort>(Vec(76, 355), module, AbcdSeq::ROOT_INPUT));
 
 	addParam(createParam<JwSmallSnapKnob>(Vec(111, paramY), module, AbcdSeq::OCTAVE_PARAM));